+++ lib/deimos/active_record_consume/batch_consumption.rb
@@ -169,7 +169,7 @@ module Deimos
                                   col_proc: col_proc,
                                   replace_associations: self.class.replace_associations,
                                   bulk_import_id_generator: self.class.bulk_import_id_generator,
-                                  backfill_associations: self.class.backfill_associations,
+                                  save_associations_first: self.class.save_associations_first,
                                   bulk_import_id_column: self.class.bulk_import_id_column)
         ActiveSupport::Notifications.instrument('batch_consumption.valid_records', {
                                                   records: updater.mass_update(record_list),