+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 - Feature: Enable `producers.persistent_connections` phobos setting
-- Feature: Add consumer configuration, `backfill_associations` to import associated records of primary class prior to upserting primary records
+- Feature: Add consumer configuration, `save_associations_first` to save associated records of primary class prior to upserting primary records. Foreign key of associated records are assigned to the record class prior to saving the record class
 
 # 1.24.2 - 2024-05-01
 - Fix: Deprecation notice with Rails 7.